+++ utils/openapiRouter.js
@@ -4,6 +4,7 @@
 const logger = require('../logger');
 const controllers = require('../controllers');
 const Services = require('../services');
+const logger = require("../logger");
 
 function handleError(err, request, response, next) {
   logger.error(err);